+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add `explicit_infinite` and `dtstart` options to `humanReadable` to respectivity omit "forever" and the start date from the sentence.
+- RFC parser will not accept multiple DTSTART or RRULE lines
 
 ## [1.4.2] - 2017-03-29